backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pymysql==1.1.1
cryptography==44.0.0
redis==5.2.1
pyjwt==2.10.1
bcrypt==4.3.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
ruff==0.9.1
